Enabled MINIMIZE_COPIES for ffi-cuda backend
