+++ hat/backends/ffi/opencl/src/main/java/hat/backend/ffi/OpenCLBackend.java
@@ -32,7 +32,7 @@ import hat.buffer.BufferTracker;
 import hat.callgraph.KernelCallGraph;
 import hat.ifacemapper.BufferState;
 
-public class OpenCLBackend extends C99FFIBackend implements BufferTracker {
+public class OpenCLBackend extends C99FFIBackend {
 
    // final Config config;
 
@@ -78,81 +78,4 @@ public class OpenCLBackend extends C99FFIBackend implements BufferTracker {
 
     }
 
-    @Override
-    public void preMutate(Buffer b) {
-        switch (b.getState()) {
-            case BufferState.NO_STATE:
-            case BufferState.NEW_STATE:
-            case BufferState.HOST_OWNED:
-            case BufferState.DEVICE_VALID_HOST_HAS_COPY: {
-                if (config.isSHOW_STATE()) {
-                    System.out.println("in preMutate state = " + b.getStateString() + " no action to take");
-                }
-                break;
-            }
-            case BufferState.DEVICE_OWNED: {
-                backendBridge.getBufferFromDeviceIfDirty(b);// calls through FFI and might block when fetching from device
-                if (config.isSHOW_STATE()) {
-                    System.out.print("in preMutate state = " + b.getStateString() + " we pulled from device ");
-                }
-                b.setState(BufferState.DEVICE_VALID_HOST_HAS_COPY);
-                if (config.isSHOW_STATE()) {
-                    System.out.println("and switched to " + b.getStateString());
-                }
-                break;
-            }
-            default:
-                throw new IllegalStateException("Not expecting this state ");
-        }
-    }
-
-    @Override
-    public void postMutate(Buffer b) {
-        if (config.isSHOW_STATE()) {
-            System.out.print("in postMutate state = " + b.getStateString() + " no action to take ");
-        }
-        if (b.getState() != BufferState.NEW_STATE) {
-            b.setState(BufferState.HOST_OWNED);
-        }
-        if (config.isSHOW_STATE()) {
-            System.out.println("and switched to (or stayed on) " + b.getStateString());
-        }
-    }
-
-    @Override
-    public void preAccess(Buffer b) {
-        switch (b.getState()) {
-            case BufferState.NO_STATE:
-            case BufferState.NEW_STATE:
-            case BufferState.HOST_OWNED:
-            case BufferState.DEVICE_VALID_HOST_HAS_COPY: {
-                if (config.isSHOW_STATE()) {
-                    System.out.println("in preAccess state = " + b.getStateString() + " no action to take");
-                }
-                break;
-            }
-            case BufferState.DEVICE_OWNED: {
-                backendBridge.getBufferFromDeviceIfDirty(b);// calls through FFI and might block when fetching from device
-
-                if (config.isSHOW_STATE()) {
-                    System.out.print("in preAccess state = " + b.getStateString() + " we pulled from device ");
-                }
-                b.setState(BufferState.DEVICE_VALID_HOST_HAS_COPY);
-                if (config.isSHOW_STATE()) {
-                    System.out.println("and switched to " + b.getStateString());
-                }
-                break;
-            }
-            default:
-                throw new IllegalStateException("Not expecting this state ");
-        }
-    }
-
-
-    @Override
-    public void postAccess(Buffer b) {
-        if (config.isSHOW_STATE()) {
-            System.out.println("in postAccess state = " + b.getStateString());
-        }
-    }
 }
